+++ src/documents/teamSheet.mjs
@@ -5,7 +5,7 @@ export class TeamSheet extends ItemSheet {
     const options = super.defaultOptions;
     return {
       ...options,
-      template: `modules/${CFG.id}/templates/item-sheet.hbs`,
+      template: `modules/${CFG.id}/templates/items/item-sheet.hbs`,
       classes: [...options.classes, "rebellion", "item", "team"],
       tabs: [
         {